Move patches to app/patches
This should make the patches in line with the Flickwerk conventions.
